requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.2
filelock==3.12.4
fsspec==2023.9.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
orjson==3.9.7
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.0
tqdm==4.66.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
